Removed unnecessary float cast at HitGround

+++ render/screens/game/ground_debug.go
@@ -38,7 +38,7 @@ func HitGround(location rl.Vector3) rl.Vector3 {
 func Hit(location rl.Vector3) float32 {
 	var bestHit rl.RayCollision
 	for _, mesh := range terrainModel.GetMeshes() {
-		ray := rl.NewRay(rl.NewVector3(float32(location.X), location.Y+0.65, float32(location.Z)), rl.NewVector3(0, -1, 0))
+		ray := rl.NewRay(rl.NewVector3(location.X, location.Y+0.65, location.Z), rl.NewVector3(0, -1, 0))
 		hit := rl.GetRayCollisionMesh(ray, mesh, terrainModel.Transform)
 		if hit.Hit {
